Morello: fix prctl
Some arguments are pointers (capabilities), depending on op.
Changing 'unsigned long' to 'uintptr_t' will work because
the way variadic arguments are placed onto the stack [1].

[1] https://github.com/ARM-software/abi-aa/blob/main/aapcs64-morello/aapcs64-morello.rst

+++ src/linux/prctl.c
@@ -5,11 +5,11 @@
 
 int prctl(int op, ...)
 {
-	unsigned long x[4];
+	uintptr_t x[4];
 	int i;
 	va_list ap;
 	va_start(ap, op);
-	for (i=0; i<4; i++) x[i] = VA_ARG_IF_IN_BOUNDS(ap, unsigned long);
+	for (i=0; i<4; i++) x[i] = VA_ARG_IF_IN_BOUNDS(ap, uintptr_t);
 	va_end(ap);
 	return syscall(SYS_prctl, op, x[0], x[1], x[2], x[3]);
 }
